requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cmake==3.22.1
colorama==0.4.4
configparser==5.2.0
cryptography==35.0.0
# dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
docker-pycreds==0.4.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.4.2
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
fvcore==0.1.5.post20210630
gitdb==4.0.9
GitPython==3.1.25
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
huggingface-hub==0.2.1
hydra-core==1.1.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
iopath==0.1.9
ipykernel==6.6.1
ipython==7.31.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
linformer==0.2.1
lmdb==1.2.1
Markdown==3.3.6
matplotlib-inline==0.1.3
msgpack==0.5.6
multidict==5.2.0
munch==2.5.0
nest-asyncio==1.5.4
Nuitka==0.6.17.3
numpy==1.19.2
oauthlib==3.1.1
omegaconf==2.1.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
portalocker==2.3.2
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
pyarmor==6.8.1
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
Pygments==2.11.1
# PyGObject==3.36.0
pyparsing==3.0.6
python-dateutil==2.8.2
# pytorch-lightning==1.4.0
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
regex==2018.1.10
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.5.1
seqeval==1.2.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.0.0
timm==0.4.12
tokenizers==0.10.3
# torch==1.9.0+cu111
# torchmetrics==0.6.2
# torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.11.1
typing_extensions==4.0.1
urllib3==1.26.7
wandb==0.12.9
wcwidth==0.2.5
Werkzeug==2.0.2
yacs==0.1.8
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
